+++ myriad-scheduler/src/main/java/com/ebay/myriad/api/ClustersResource.java
@@ -119,9 +119,9 @@ public class ClustersResource {
       // Validate profile request
       Response.ResponseBuilder response = Response.status(Response.Status.ACCEPTED);
       
-      if (cfg.getServiceConfiguration(profile) != null) {
+      if (cfg.getServiceConfiguration(profile) == null) {
         response.status(Response.Status.BAD_REQUEST)
-                .entity("Sevrice does not exist: " + profile);
+                .entity("Service does not exist: " + profile);
         LOGGER.error("Provided service does not exist " + profile);
         return response.build();
       }
